# Conflicts:
#	server/src/main/java/project/server/domain/user/UserController.java
#	server/src/main/java/project/server/domain/user/UserService.java
#	server/src/main/java/project/server/domain/user/entity/User.java

+++ server/src/main/java/project/server/domain/comment/service/CommentService.java
@@ -4,23 +4,26 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.server.domain.cocktail.entity.Cocktail;
 import project.server.domain.cocktail.service.CocktailReadService;
+import project.server.domain.comment.CommentSerializer;
 import project.server.domain.comment.dto.CommentDto;
 import project.server.domain.comment.entity.Comment;
 import project.server.domain.comment.repository.CommentRepository;
-import project.server.domain.user.User;
-import project.server.domain.user.UserService;
-import project.server.exception.BusinessLogicException;
-import project.server.exception.ExceptionCode;
+import project.server.domain.user.entity.User;
+import project.server.domain.user.service.UserService;
+import project.server.global.exception.BusinessLogicException;
+import project.server.global.exception.ExceptionCode;
 
 @Service
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final CommentSerializer commentSerializer;
     private final CocktailReadService cocktailReadService;
     private final UserService userService;
 
     public CommentService(CommentRepository commentRepository,
-                          CocktailReadService cocktailReadService, UserService userService) {
+                          CommentSerializer commentSerializer, CocktailReadService cocktailReadService, UserService userService) {
         this.commentRepository = commentRepository;
+        this.commentSerializer = commentSerializer;
         this.cocktailReadService = cocktailReadService;
         this.userService = userService;
     }
@@ -30,33 +33,44 @@ public class CommentService {
         User user = userService.findUserByEmail(email);
         Cocktail cocktail = cocktailReadService.readCocktail(cocktailId);
         Comment comment = post.postToEntity();
-        comment.setUserName(user.getName());
-        comment.setUserId(user.getUserId());
-        comment.setCocktail(cocktail);
+        comment.setUser(user);
         Comment savedComment = commentRepository.save(comment);
-        return savedComment.entityToResponse();
+        cocktail.addComment(savedComment);
+        return commentSerializer.entityToResponse(savedComment);
     }
 
     @Transactional(readOnly = true)
     public CommentDto.Response readComment(long commentId) {
         Comment comment = findCommentById(commentId);
-        return comment.entityToResponse();
+        return commentSerializer.entityToResponse(comment);
     }
 
-    public CommentDto.Response updateComment(Long commentId, CommentDto.Patch patch) {
+    @Transactional
+    public CommentDto.Response updateComment(String email, long commentId, CommentDto.Patch patch) {
+        User user = userService.findUserByEmail(email);
         Comment comment = findCommentById(commentId);
+        if(!user.hasAuthority(comment.getUserId())){
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);
+        }
         comment.setContent(patch.getContent());
-        Comment savedComment = commentRepository.save(comment);
-        return savedComment.entityToResponse();
+        return commentSerializer.entityToResponse(comment);
     }
 
+    @Transactional
+    public void deleteComment(String email, long commentId, long cocktailId) {
+        User user = userService.findUserByEmail(email);
+        Comment comment = findCommentById(commentId);
+        if(!user.hasAuthority(comment.getUserId())){
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);
+        }
+        Cocktail cocktail = cocktailReadService.readCocktail(cocktailId);
+        cocktail.removeComment(comment);
+        commentRepository.delete(comment);
+    }
+
+    @Transactional(readOnly = true)
     public Comment findCommentById(long commentId) {
         return commentRepository.findById(commentId).orElseThrow(() ->
                 new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND));
-    }
-
-    public void deleteComment(long commentId) {
-        Comment comment = findCommentById(commentId);
-        commentRepository.delete(comment);
     }
 }
